agents/requirements.txt
crewai>=0.80.0
crewai-tools>=0.14.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
